+++ src/commands/builders/applicationSetup.ts
@@ -1,5 +1,6 @@
 import { ChannelType, PermissionsBitField, SlashCommandBuilder } from 'discord.js';
 import { lang } from '../../utils';
+import { createForumChannelOption, createTextChannelOption } from './factories';
 
 const tl = lang.application.setup;
 
@@ -8,19 +9,9 @@ export const applicationSetup = new SlashCommandBuilder()
   .setName('application-setup')
   .setDescription(tl.cmdDescrp)
   .setDefaultMemberPermissions(PermissionsBitField.Flags.Administrator)
+  .addChannelOption(option => createTextChannelOption(option, { description: tl.options.channel, required: false }))
   .addChannelOption(option =>
-    option
-      .setName('channel')
-      .setDescription(tl.options.channel)
-      .setRequired(false)
-      .addChannelTypes(ChannelType.GuildText),
-  )
-  .addChannelOption(option =>
-    option
-      .setName('archive')
-      .setDescription(tl.options.archive)
-      .setRequired(false)
-      .addChannelTypes(ChannelType.GuildForum),
+    createForumChannelOption(option, { name: 'archive', description: tl.options.archive, required: false }),
   )
   .addChannelOption(option =>
     option